Add numbers to grammer list
